refactor: styling restructure
