+++ frontend/src/index.tsx
@@ -4,7 +4,10 @@ import App from "./App";
 
 import { ThemeProvider } from "@mui/material/styles";
 import CssBaseline from "@mui/material/CssBaseline";
-import theme from "./theme";
+
+// Styling Imports
+import theme from "./styles/theme";
+import './styles/main.scss'; 
 
 // Create React root, ensuring that root element is present
 ReactDOM.createRoot(document.getElementById("root")!).render(
